BE/pyproject.toml
[project]
name = "janus-agent"
version = "0.1.0"
description = "Janus AI Agent Backend"
dependencies = [
    "flask==3.0.2",
    "flask-cors==4.0.0",
    "flask-sqlalchemy==3.1.1",
    "psycopg2-binary==2.9.9",
    "langchain==0.1.13",
    "langchain-openai==0.1.1",
    "langchain-pinecone==0.0.3",
    "langgraph==0.0.30",
    "pinecone-client==3.2.2",
    "python-dotenv==1.0.1",
    "gunicorn==21.2.0",
    "requests==2.31.0",
    "google-api-python-client==2.122.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-flask>=1.3.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
